+++ twcore/twcore/core/FlagPosition.java
@@ -2,31 +2,31 @@ package twcore.core;
 
 public class FlagPosition extends SubspaceEvent {
     
-    private int m_flagID;
-    private int m_xLocation;
-    private int m_yLocation;
-    private int m_team;
+    private short m_flagID;
+    private short m_xLocation;
+    private short m_yLocation;
+    private short m_team;
     
     public FlagPosition( ByteArray array ) {
-        m_flagID = (int)array.readLittleEndianShort( 1 );
-        m_xLocation = (int)array.readLittleEndianShort( 3 );
-        m_yLocation = (int)array.readLittleEndianShort( 5 );
-        m_team = (int)array.readLittleEndianShort( 7 );
+        m_flagID = array.readLittleEndianShort( 1 );
+        m_xLocation = array.readLittleEndianShort( 3 );
+        m_yLocation = array.readLittleEndianShort( 5 );
+        m_team = array.readLittleEndianShort( 7 );
     }
     
-    public int getFlagID() {
+    public short getFlagID() {
         return m_flagID;
     }
     
-    public int getXLocation() {
+    public short getXLocation() {
         return m_xLocation;
     }
     
-    public int getYLocation() {
+    public short getYLocation() {
         return m_yLocation;
     }
     
-    public int getTeam() {
+    public short getTeam() {
         return m_team;
     }
 }
