+++ twcore/twcore/core/TurfFlagUpdate.java
@@ -3,12 +3,12 @@ package twcore.core;
 public class TurfFlagUpdate extends SubspaceEvent {
 
     int         m_flagId;
-    int         m_frequency;
+    short       m_frequency;
     boolean     m_claimed = false;
 
     public TurfFlagUpdate( ByteArray array, int flagId ){
         m_flagId = flagId;
-        m_frequency = (int)array.readLittleEndianShort( 1 );
+        m_frequency = array.readLittleEndianShort( 1 );
         if( m_frequency > -1 ){
             m_claimed = true;
         }
@@ -18,7 +18,7 @@ public class TurfFlagUpdate extends SubspaceEvent {
         return m_flagId;
     }
 
-    public int getFrequency(){
+    public short getFrequency(){
         return m_frequency;
     }
